v1.2.0 リリース (#98)
* Update for display number

- Black number
- Blue number

* Update BoxGizmoDrawer.cs

Update Black Number position adjust

* fix text draw go out of rectangle when scale

* re-fix text go out rectangle when scale screen on window

* Update for meshcode sellected

* Update to fix comment at 1403

* Update to refactor scripts

Update to refactor scripts

* fix lue color, zoom in

* Update to finish refactor scripts Display Number

Update to finish refactor scripts Display Number

* Update for refactor comment

Update for refactor comment

* Update for UNITY_EDITOR

Update for UNITY_EDITOR

* Remove Debug Error Log

* Update to refactor : Remove unnecessary code

* fix mesh code text position

* libplateauをリリースビルドに、libplateauのライセンス変更を反映

* gmlの並列処理に関するエラーを修正、demのインポートができないバグを修正 (#97)

* gmlインポートをfetchとimportに分割

* gmlの並列処理に関するエラーを改善

* demのインポートが異常に重かったバグを修正

* 微修正

* dllをリリースビルドに

---------

+++ Runtime/CityImport/Setting/CityLoadConfig.cs
@@ -70,8 +70,8 @@ namespace PLATEAU.CityImport.Setting
                 var package = pair.Key;
                 var maxLod = pair.Value;
                 var predefined = CityModelPackageInfo.GetPredefined(package);
-                var val = new PackageLoadSetting(true, predefined.hasAppearance, (uint)predefined.minLOD,
-                    (uint)maxLod, MeshGranularity.PerPrimaryFeatureObject, false);
+                var val = new PackageLoadSetting(true, predefined.hasAppearance, predefined.minLOD,
+                    maxLod, MeshGranularity.PerPrimaryFeatureObject, false);
                 this.perPackagePairSettings.Add(package, val);
             }
         }
